requirements.txt
setuptools==57.4.0
elasticsearch==7.12.0
netaddr==0.8.0
flask==2.0.1
terminable-thread==0.7.1 # library_name=terminable_thread
watchdog==2.1.3
pyshark==0.4.3
oschmod==0.3.12
argparse==1.4.0
PyYAML==5.4.1 # library_name=yaml
flask-swagger==0.2.14 # library_name=flask_swagger
flask-swagger-ui==3.36.0 # library_name=flask_swagger_ui